requirements.txt
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0
matplotlib>=3.7.0
plotly>=5.17.0
astropy>=5.3.0
requests>=2.31.0
pyyaml>=6.0
click>=8.1.0
Pillow>=10.0.0
